♻️ refactor: 헤더를 createHeader 컴포넌트로 리팩토링
- 헤더 생성 로직을 Header.js의 재사용 가능한 createHeader 컴포넌트로 이동
